### 2022-06-26 #4511 @editorialbot
[PRE REVIEW]: solid_dmft: gray-boxing DFT+DMFT materials simulations with TRIQS
**Submitting author:** @the-hampel (<a href="http://orcid.org/0000-0003-1041-8614">Alexander Hampel</a>)
**Repository:** https://github.com/TRIQS/solid_dmft
**Branch with paper.md** (empty if default branch): joss
**Version:** 3.1.1
