+++ paper/paper.bib
@@ -30,7 +30,7 @@ volume = {244},
 pages = {277-294},
 year = {2019},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2019.07.003},
+doi = {doi.org/10.1016/j.cpc.2019.07.003},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465519302140},
 author = {Sangkook Choi and Patrick Semon and Byungkyun Kang and Andrey Kutepov and Gabriel Kotliar},
 keywords = {Correlated electron system, First principles, Dynamical mean-field theory, Electronic structure}
@@ -137,7 +137,7 @@ eprint = {https://doi.org/10.1063/1.5144261}
   doi = {10.1103/PhysRevB.104.035102},
 }
 
-﻿@Article{Hampel:2019,
+@Article{Hampel:2019,
 author={Hampel, Alexander
 and Liu, Peitao
 and Franchini, Cesare
