### 2022-07-04 d834a2f
adapt to sumk mesh changes in dfttools
- initialize mesh first in dmft_cycle
- use sum_k.mesh in all instances
- remove iw_or_w
