+++ python/solid_dmft/dmft_tools/solver.py
@@ -24,7 +24,7 @@
 import numpy as np
 from itertools import product
 
-from triqs.gf import GfImTime, GfReTime, GfImFreq, GfReFreq, GfLegendre, BlockGf, make_hermitian, Omega, iOmega_n, make_gf_from_fourier, fit_hermitian_tail
+from triqs.gf import MeshImTime, MeshReTime, MeshReFreq, MeshLegendre, Gf, BlockGf, make_hermitian, Omega, iOmega_n, make_gf_from_fourier, fit_hermitian_tail
 from triqs.gf.tools import inverse, make_zero_tail
 from triqs.gf.descriptors import Fourier
 from triqs.operators import c_dag, c, Operator
@@ -195,8 +195,8 @@ class SolverStructure:
 
         # create all ImFreq instances
         self.n_iw = self.general_params['n_iw']
-        self.G_freq = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=GfImFreq, space='solver',
-                                                           beta=self.general_params['beta'], n_points=self.n_iw)
+        self.G_freq = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=Gf, space='solver',
+                                                           mesh=self.sum_k.mesh)
         # copy
         self.Sigma_freq = self.G_freq.copy()
         self.G0_freq = self.G_freq.copy()
@@ -205,8 +205,10 @@ class SolverStructure:
 
         # create all ImTime instances
         self.n_tau = self.general_params['n_tau']
-        self.G_time = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=GfImTime, space='solver',
-                                                           beta=self.general_params['beta'], n_points=self.n_tau)
+        self.G_time = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=Gf, space='solver',
+                                                           mesh=MeshImTime(beta=self.general_params['beta'],
+                                                                           S='Fermion', n_tau=self.n_tau)
+                                                           )
         # copy
         self.Delta_time = self.G_time.copy()
 
@@ -218,8 +220,10 @@ class SolverStructure:
             or self.general_params['solver_type'] == 'hubbardI' and self.solver_params['measure_G_l']):
 
             self.n_l = self.general_params['n_l']
-            self.G_l = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=GfLegendre, space='solver',
-                                                            beta=self.general_params['beta'], n_points=self.n_l)
+            self.G_l = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=Gf, space='solver',
+                                                            mesh=MeshLegendre(beta=self.general_params['beta'],
+                                                                              max_n=self.n_l, S='Fermion')
+                                                            )
             # move original G_freq to G_freq_orig
             self.G_time_orig = self.G_time.copy()
 
@@ -237,8 +241,8 @@ class SolverStructure:
 
         # create all ReFreq instances
         self.n_w = self.general_params['n_w']
-        self.G_freq = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=GfReFreq, space='solver',
-                                                           n_points=self.n_w, window = self.general_params['w_range'])
+        self.G_freq = self.sum_k.block_structure.create_gf(ish=self.icrsh, gf_function=Gf, space='solver',
+                                                           mesh=self.sum_k.mesh)
         # copy
         self.Sigma_freq = self.G_freq.copy()
         self.G0_freq = self.G_freq.copy()
@@ -248,7 +252,7 @@ class SolverStructure:
[…]
