+++ _sources/input_output/DMFT_input/general.rst.txt
@@ -1,0 +1,488 @@
+[general]: General parameters
+-----------------------------
+
+Includes the majority of the parameters
+
+
+
+
+
+
+.. admonition:: seedname 
+ 	:class: intag  
+ 
+            **type=** str or list of str
+
+            seedname for h5 archive or for multiple if calculations should be connected
+
+.. admonition:: jobname 
+ 	:class: intag  
+ 
+            **type=** str or list of str;  **optional**;  **default=** seedname
+
+            one or multiple jobnames specifying the output directories
+
+.. admonition:: csc 
+ 	:class: intag  
+ 
+            **type=** bool;  **optional**;  **default=** False
+
+            are we doing a CSC calculation?
+
+.. admonition:: plo_cfg 
+ 	:class: intag  
+ 
+            **type=** str;  **optional**;  **default=** 'plo.cfg'
+
+            config file for PLOs for the converter
+
+.. admonition:: h_int_type 
+ 	:class: intag  
+ 
+            **type=** string
+
+            interaction type:
+
+            * density_density: used for full d-shell or eg- or t2g-subset
+            * kanamori: only physical for the t2g or the eg subset
+            * full_slater: used for full d-shell or eg- or t2g-subset
+            * crpa: use the cRPA matrix as interaction Hamiltonian
+            * crpa_density_density: use the density-density terms of the cRPA matrix
+            * dynamic: use dynamic U from h5 archive
+
+            Needs to be stored as Matsubara Gf under dynamic_U/U_iw in the input h5
+
+.. admonition:: U 
+ 	:class: intag  
+ 
+            **type=** float or comma separated list of floats
+
+            U values for impurities if only one value is given, the same U is assumed for all impurities
+
+.. admonition:: J 
+ 	:class: intag  
+ 
+            **type=** float or comma separated list of floats
+
+            J values for impurities if only one value is given, the same J is assumed for all impurities
+
+.. admonition:: ratio_F4_F2 
+ 	:class: intag  
+ 
+            **type=** float or comma separated list of floats;  **optional**;  **default=** 'none'
+
+            Ratio between the Slater integrals  F_4 and F_2. Only used for the
+            interaction Hamiltonians 'density_density' and 'full_slater' and
+            only for d-shell impurities, where the default is 0.63.
+
+.. admonition:: beta 
+ 	:class: intag  
+ 
+            **type=** float, only used if solver ImFreq
+
+            inverse temperature for Greens function etc
+
+.. admonition:: n_iter_dmft_first 
+ 	:class: intag  
+ 
+            **type=** int;  **optional**;  **default=**  10
+
+            number of iterations in first dmft cycle to converge dmft solution
+
+.. admonition:: n_iter_dmft_per 
+ 	:class: intag  
+ 
+            **type=** int;  **optional**;  **default=**  2
+
+            number of iterations per dmft step in CSC calculations
+
+.. admonition:: n_iter_dmft 
+ 	:class: intag  
+ 
+            **type=** int
+
+            number of iterations per dmft cycle after first cycle
+
+.. admonition:: dc_type 
+ 	:class: intag  
+ 
+            **type=** int
+
+            Type of double counting correction considered:
+            * 0: FLL
+            * 1: held formula, needs to be used with slater-kanamori h_int_type=2
+            * 2: AMF
+            * 3: FLL for eg orbitals only with U,J for Kanamori
+
+.. admonition:: prec_mu 
+ 	:class: intag  
+ 
+            **type=** float
+
+            general precision for determining the chemical potential at any time calc_mu is called
+
+.. admonition:: dc_dmft 
[…]
