+++ _sources/_ref/dmft_tools.legendre_filter.rst.txt
@@ -1,0 +1,21 @@
+dmft\_tools.legendre\_filter
+============================
+
+.. automodule:: dmft_tools.legendre_filter
+    :members:
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
