+++ _sources/_ref/dmft_tools.formatter.rst.txt
@@ -1,0 +1,21 @@
+dmft\_tools.formatter
+=====================
+
+.. automodule:: dmft_tools.formatter
+    :members:
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
