+++ _sources/_ref/util.rst.txt
@@ -1,0 +1,34 @@
+﻿util
+====
+
+.. automodule:: util
+    :members:
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+    :toctree:
+    :template: autosummary_module_template.rst 
+    :recursive:
+
+    
+    util.symmetrize_gamma_file
+    util.update_dmft_config
+    util.update_results_h5
+    util.write_kslice_to_h5
+
+
