+++ requirements.txt
@@ -1,1 +1,2 @@
 # Required python packages for this application (these should also be added to Dockerfile for Jenkins)
+numpy scipy argparse pytest scikit-image
