### 2022-07-07 e1a4753
fix load Sigma with new gf_struct in triqs 3.1.x
