### 2022-07-07 4d4cada
fix load Sigma with new gf_struct in triqs 3.1.x
