### 2022-07-14 20f8608
fix bug in max_G_diff function ignoring norm_temp
