### 2022-07-25 #4511 @xwang862
Hi Rachel, thanks for your invitation. I am not sure if I am eligible to review this work because of a Conflict of Interest - I worked in the same institute as two of the authors in the past few years.

> On Jul 24, 2022, at 5:41 PM, Rachel Kurchin ***@***.***> wrote:
>
>
> 👋 @shivupa <https://github.com/shivupa> and @xwang862 <https://github.com/xwang862>, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html <https://joss.readthedocs.io/en/latest/review_criteria.html>
