### 2022-07-28 ba632aa
Update PCB module
* start to add docstrings to functions
* add various checks / asserts of input parameters
* calc_kslice supports now band mode
* streamline chemical potential interface: get_dmft_bands does not
  return dft_mu anymore, rename dft_mu -> tb_mu. tb_mu is subtracted
  from TB model and can be later added to the spectral function
  again when DMFT has been run with removed tb_mu.
* additional parameter mu_shift can be used to tune the final mu for the
  calculation of the spectral function
* when calculating fermi slices tb_data holds now the full 2D kpoint
  mesh correctly
* eige
[…]
