+++ python/solid_dmft/postprocessing/plot_correlated_bands.py
@@ -51,11 +51,11 @@ from triqs.lattice.utils import TB_from_wannier90, k_space_path
 
 def _linefit(x, y, interval, spacing=50, addspace=0.0):
 
-    calc_Z = lambda slope: 1/(1-slope)
+    def calc_Z(slope): return 1/(1-slope)
 
     x = np.array(x)
     lim_l, lim_r = interval
-    indices = np.where(np.logical_and(x>=lim_l, x<=lim_r))
+    indices = np.where(np.logical_and(x >= lim_l, x <= lim_r))
     fit = np.polyfit(x[indices], y[indices], 1)
     slope = fit[1]
     Z = calc_Z(slope)
@@ -64,29 +64,32 @@ def _linefit(x, y, interval, spacing=50, addspace=0.0):
 
     return x_cont, f_x(x_cont), fit
 
+
 def lambda_matrix_w90_t2g(add_lambda):
 
     lambda_x, lambda_y, lambda_z = add_lambda
 
-    lambda_matrix = np.zeros((6,6), dtype=complex)
-    lambda_matrix[0,1] = -1j*lambda_z/2.0
-    lambda_matrix[0,5] =  1j*lambda_x/2.0
-    lambda_matrix[1,5] =    -lambda_y/2.0
-    lambda_matrix[2,3] = -1j*lambda_x/2.0
-    lambda_matrix[2,4] =     lambda_y/2.0
-    lambda_matrix[3,4] =  1j*lambda_z/2.0
+    lambda_matrix = np.zeros((6, 6), dtype=complex)
+    lambda_matrix[0, 1] = -1j*lambda_z/2.0
+    lambda_matrix[0, 5] = 1j*lambda_x/2.0
+    lambda_matrix[1, 5] = -lambda_y/2.0
+    lambda_matrix[2, 3] = -1j*lambda_x/2.0
+    lambda_matrix[2, 4] = lambda_y/2.0
+    lambda_matrix[3, 4] = 1j*lambda_z/2.0
     lambda_matrix += np.transpose(np.conjugate(lambda_matrix))
 
     return lambda_matrix
+
 
 def change_basis(n_orb, orbital_order_to, orbital_order_from):
 
     change_of_basis = np.eye(n_orb)
     for ct, orb in enumerate(orbital_order_to):
         orb_idx = orbital_order_from.index(orb)
-        change_of_basis[orb_idx,:] = np.roll(np.eye(n_orb,1),ct)[:,0]
+        change_of_basis[orb_idx, :] = np.roll(np.eye(n_orb, 1), ct)[:, 0]
 
     return change_of_basis
+
 
 def print_matrix(matrix, n_orb, text):
 
@@ -101,12 +104,13 @@ def print_matrix(matrix, n_orb, text):
     for row in matrix:
         print((' '*4 + fmt).format(*row))
 
+
 def _sigma_from_dmft(n_orb, orbital_order, with_sigma, spin, block, orbital_order_dmft, eta=0.0, **specs):
 
     if with_sigma == 'calc':
         print('Setting Sigma from {}'.format(specs['dmft_path']))
 
-        with HDFArchive(specs['dmft_path'],'r') as ar:
+        with HDFArchive(specs['dmft_path'], 'r') as ar:
             try:
                 sigma = ar['DMFT_results'][specs['it']]['Sigma_freq_0']
                 assert isinstance(sigma.mesh, MeshReFreq), 'Imported Greens function must be real frequency'
@@ -115,21 +119,21 @@ def _sigma_from_dmft(n_orb, orbital_order, with_sigma, spin, block, orbital_orde
                     sigma = ar['DMFT_results'][specs['it']]['Sigma_maxent_0']
                 except KeyError:
[…]
