### 2022-07-28 c8976e2
resolve suggestions from @phibeck and update tutorial
