+++ python/solid_dmft/postprocessing/plot_correlated_bands.py
@@ -120,14 +120,14 @@ def _sigma_from_dmft(n_orb, orbital_order, with_sigma, spin, block, orbital_orde
                 except KeyError:
                     raise KeyError('Provide either "Sigma_freq_0" in real frequency or "Sigma_maxent_0".')
             dc = ar['DMFT_results'][specs['it']]['DC_pot'][0][spin][0, 0]
-            mu = ar['DMFT_results'][specs['it']]['chemical_potential_post']
+            mu_dmft = ar['DMFT_results'][specs['it']]['chemical_potential_post']
 
     else:
         print('Setting Sigma from memory')
 
         sigma = with_sigma
         dc = specs['dc'][0][spin][0, 0]
-        mu = specs['dmft_mu']
+        mu_dmft = specs['mu_dmft']
 
     block_spin = spin + '_' + str(block) if with_sigma == 'calc' else spin
     SOC = (spin == 'ud')
@@ -175,7 +175,7 @@ def _sigma_from_dmft(n_orb, orbital_order, with_sigma, spin, block, orbital_orde
                 continue
             sigma_interpolated[ct1, ct2] = interpolate_sigma(w_mesh, w_mesh_dmft, ct1, ct2)
 
-    return sigma_interpolated, mu, freq_dict
+    return sigma_interpolated, mu_dmft, freq_dict
 
 
 def _sigma_from_model(n_orb, orbital_order, zeroth_order, first_order, eta=0.0, **w):
@@ -198,16 +198,16 @@ def _sigma_from_model(n_orb, orbital_order, zeroth_order, first_order, eta=0.0, 
     return sigma_interpolated, freq_dict
 
 
-def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=np.array([None]),
-                proj_knu=np.array([None]), trace=True, **freq_dict):
+def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
+                proj_nuk=None, trace=True, **freq_dict):
     '''
-    calculate lattice spectral function for given TB dispersion / e_mat and self-energy
+    calculate slice of lattice spectral function for given TB dispersion / e_mat and self-energy
 
     Parameters
     ----------
     n_orb : int
           number of Wannier orbitals
-    proj_knu : optinal, 2D numpy array (n_k, n_orb)
+    proj_nuk : optinal, 2D numpy array (n_k, n_orb)
           projections to be applied on A(k,w) in band basis. Only works when band_basis=True
 
     Returns
@@ -221,7 +221,7 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=np.array([N
     def upscale(quantity, n_orb): return quantity * np.identity(n_orb)
     mu = upscale(mu, n_orb)
     eta = upscale(eta, n_orb)
-    if not e_vecs.any() is None:
+    if isinstance(e_vecs, np.ndarray):
         sigma_rot = np.zeros(sigma.shape, dtype=complex)
 
     w_vec = np.array([upscale(freq_dict['w_mesh'][w], n_orb) for w in range(freq_dict['n_w'])])
@@ -233,11 +233,11 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=np.array([N
         else:
             alatt_k_w = np.zeros((n_k, freq_dict['n_w'], n_orb))
 
-        def invert_and_trace(w, eta, mu, e_mat, sigma, trace, proj=np.array([None])):
+        def invert_and_trace(w, eta, mu, e_mat, sigma, trace, proj=None):
[…]
