+++ paper/paper.bib
@@ -30,7 +30,7 @@ volume = {244},
 pages = {277-294},
 year = {2019},
 issn = {0010-4655},
-doi = {doi.org/10.1016/j.cpc.2019.07.003},
+doi = {10.1016/j.cpc.2019.07.003},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465519302140},
 author = {Sangkook Choi and Patrick Semon and Byungkyun Kang and Andrey Kutepov and Gabriel Kotliar},
 keywords = {Correlated electron system, First principles, Dynamical mean-field theory, Electronic structure}
