### 2022-07-29 #4623 @the-hampel
> @the-hampel, do fix that one flagged reference whenever you have a moment :)

Hi @rkurchin,
I just fixed the one DOI that you pointed out: https://github.com/TRIQS/solid_dmft/commit/fe459150bbb4ca6e3815eec1fd8c5f33c98b14aa
