### 2022-07-29 bac1faf
add proj_on_orb functionality to Akw
