+++ python/solid_dmft/postprocessing/plot_correlated_bands.py
@@ -199,7 +199,7 @@ def _sigma_from_model(n_orb, orbital_order, zeroth_order, first_order, eta=0.0, 
 
 
 def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
-                proj_nuk=None, trace=True, **freq_dict):
+                proj_nuk=None, trace=True, orb_proj=None, **freq_dict):
     '''
     calculate slice of lattice spectral function for given TB dispersion / e_mat and self-energy
 
@@ -223,6 +223,12 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
     eta = upscale(eta, n_orb)
     if isinstance(e_vecs, np.ndarray):
         sigma_rot = np.zeros(sigma.shape, dtype=complex)
+
+        # now we merge proj_nuk and orb_proj (has reverse shape)
+        if isinstance(proj_nuk, np.ndarray) and isinstance(orb_proj, np.ndarray):
+            proj_nuk = proj_nuk * orb_proj.transpose()
+        elif not isinstance(proj_nuk, np.ndarray) and isinstance(orb_proj, np.ndarray):
+            proj_nuk = orb_proj.transpose()
 
     w_vec = np.array([upscale(freq_dict['w_mesh'][w], n_orb) for w in range(freq_dict['n_w'])])
     n_k = e_mat.shape[2]
@@ -278,7 +284,7 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
     return alatt_k_w
 
 
-def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=None, **freq_dict):
+def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=None, orb_proj=None, **freq_dict):
     '''
     calculate lattice spectral function for given TB dispersion / e_mat and self-energy
 
@@ -307,6 +313,12 @@ def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=N
     if isinstance(e_vecs, np.ndarray):
         sigma_rot = np.zeros(sigma.shape, dtype=complex)
 
+        # now we merge proj_nuk and orb_proj (has reverse shape)
+        if isinstance(proj_nuk, np.ndarray) and isinstance(orb_proj, np.ndarray):
+            proj_nuk = proj_nuk * orb_proj.transpose()
+        elif not isinstance(proj_nuk, np.ndarray) and isinstance(orb_proj, np.ndarray):
+            proj_nuk = orb_proj.transpose()
+
     n_kx, n_ky = e_mat.shape[2:4]
 
     if not qp_bands:
@@ -316,7 +328,7 @@ def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=N
             # inversion is automatically vectorized over first axis of 3D array (omega first index now)
             Glatt = np.linalg.inv(w + eta + mu - e_mat - sigma)
             A_nu = -1.0/np.pi * np.diagonal(Glatt).imag
-            if isinstance(proj_nuk, np.ndarray):
+            if isinstance(proj, np.ndarray):
                 A_nu = A_nu * proj
             return np.sum(A_nu)
 
@@ -363,7 +375,7 @@ def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=N
     return alatt_k_w
 
 
-def _get_tb_bands(e_mat, **specs):
+def _get_tb_bands(e_mat, proj_on_orb=[None], **specs):
     '''
[…]
