### 2022-07-30 2246b9e
add new test for pcb get_dmft_bands function
