+++ python/solid_dmft/postprocessing/plot_correlated_bands.py
@@ -199,7 +199,7 @@ def _sigma_from_model(n_orb, orbital_order, zeroth_order, first_order, eta=0.0, 
 
 
 def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
-                proj_nuk=None, trace=True, orb_proj=None, **freq_dict):
+                proj_nuk=None, trace=True, **freq_dict):
     '''
     calculate slice of lattice spectral function for given TB dispersion / e_mat and self-energy
 
@@ -207,7 +207,7 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
     ----------
     n_orb : int
           number of Wannier orbitals
-    proj_nuk : optinal, 2D numpy array (n_k, n_orb)
+    proj_nuk : optinal, 2D numpy array (n_orb, n_k)
           projections to be applied on A(k,w) in band basis. Only works when band_basis=True
 
     Returns
@@ -223,12 +223,6 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
     eta = upscale(eta, n_orb)
     if isinstance(e_vecs, np.ndarray):
         sigma_rot = np.zeros(sigma.shape, dtype=complex)
-
-        # now we merge proj_nuk and orb_proj (has reverse shape)
-        if isinstance(proj_nuk, np.ndarray) and isinstance(orb_proj, np.ndarray):
-            proj_nuk = proj_nuk * orb_proj.transpose()
-        elif not isinstance(proj_nuk, np.ndarray) and isinstance(orb_proj, np.ndarray):
-            proj_nuk = orb_proj.transpose()
 
     w_vec = np.array([upscale(freq_dict['w_mesh'][w], n_orb) for w in range(freq_dict['n_w'])])
     n_k = e_mat.shape[2]
@@ -255,7 +249,7 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
             if isinstance(e_vecs, np.ndarray):
                 sigma_rot = np.einsum('ij,jkw->ikw', e_vecs[:, :, ik].conjugate().transpose(), np.einsum('ijw,jk->ikw', sigma, e_vecs[:, :, ik]))
                 if isinstance(proj_nuk, np.ndarray):
-                    alatt_k_w[ik, :] = invert_and_trace(w_vec, eta, mu, e_mat[:, :, ik], sigma_rot, trace, proj_nuk[ik])
+                    alatt_k_w[ik, :] = invert_and_trace(w_vec, eta, mu, e_mat[:, :, ik], sigma_rot, trace, proj_nuk[:, ik])
                 else:
                     alatt_k_w[ik, :] = invert_and_trace(w_vec, eta, mu, e_mat[:, :, ik], sigma_rot, trace)
             else:
@@ -284,7 +278,7 @@ def _calc_alatt(n_orb, mu, eta, e_mat, sigma, qp_bands=False, e_vecs=None,
     return alatt_k_w
 
 
-def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=None, orb_proj=None, **freq_dict):
+def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=None, **freq_dict):
     '''
     calculate lattice spectral function for given TB dispersion / e_mat and self-energy
 
@@ -292,7 +286,7 @@ def _calc_kslice(n_orb, mu, eta, e_mat, sigma, qp_bands, e_vecs=None, proj_nuk=N
[…]
