### 2022-08-09 89ceabf
print optimal alpha in other maxent scripts
