### 2022-08-16 1478b3a
enforce PEP standard in interaction Hamiltonian
