### 2022-08-16 3dcb25b
allow different interaction Hamiltonians per impurity
