+++ test/python/test_read_config.py
@@ -102,7 +102,7 @@ def test_config_file_1():
                                                         'g0_mix_type' : 'linear',
                                                         'dc_type': 0, 'load_sigma': False, 'n_iw': 1025,
                                                         'noise_level_initial_sigma': 0.,
-                                                        'h_int_type': 'density_density',
+                                                        'h_int_type': [u'density_density'],
                                                         'mu_gap_gb2_threshold': 'none'},
                                             'advanced': {'dc_fixed_value': 'none', 'dc_fixed_occ': 'none',
                                                          'dc_nominal': False,
@@ -224,7 +224,7 @@ def test_config_file_2():
                                                         'magmom': [1.0, -1.0], 'dc_type': 0, 'n_tau': 10001,
                                                         'load_sigma': False, 'n_l': 35, 'n_iw': 1025,
                                                         'noise_level_initial_sigma': 0., 'n_iter_dmft': 6,
-                                                        'h_int_type': 'kanamori',
+                                                        'h_int_type': [u'kanamori'], 'ratio_F4_F2' : [u'none'],
                                                         'mu_gap_gb2_threshold': 'none'},
                                             'advanced': {'dc_fixed_value': 'none', 'dc_fixed_occ': 'none',
                                                          'dc_nominal': False,
