+++ python/solid_dmft/dmft_tools/interaction_hamiltonian.py
@@ -172,50 +172,47 @@ def _adapt_U_4index_for_SO(Umat_full):
     return Umat_full_SO
 
 
-def _construct_kanamori(sum_k, general_params):
+def _construct_kanamori(sum_k, general_params, icrsh):
     """
     Constructs the Kanamori interaction Hamiltonian. Only Kanamori does not
     need the full four-index matrix. Therefore, we can construct it directly
     from the parameters U and J.
     """
 
-    h_int = [None] * sum_k.n_inequiv_shells
+    # ish points to the shell representative of the current group
+    ish = sum_k.inequiv_to_corr[icrsh]
+    orb_names = list(range(sum_k.corr_shells[ish]['dim']))
+    if sum_k.SO == 0:
+        n_orb = sum_k.corr_shells[ish]['dim']
+    else:
+        assert sum_k.corr_shells[ish]['dim'] % 2 == 0
+        n_orb = sum_k.corr_shells[ish]['dim'] // 2
 
-    for icrsh in range(sum_k.n_inequiv_shells):
-        # ish points to the shell representative of the current group
-        ish = sum_k.inequiv_to_corr[icrsh]
-        orb_names = list(range(sum_k.corr_shells[ish]['dim']))
-        if sum_k.SO == 0:
-            n_orb = sum_k.corr_shells[ish]['dim']
+    if n_orb not in (2, 3):
+        mpi.report('warning: are you sure you want to use the kanamori hamiltonian '
+                   + 'outside the t2g or eg manifold?')
+
+    if general_params['solver_type'] == 'ftps':
+        # 1-band modell requires J and U' equals zero
+        if n_orb == 1:
+            up, j = 0.0, 0.0
         else:
-            assert sum_k.corr_shells[ish]['dim'] % 2 == 0
-            n_orb = sum_k.corr_shells[ish]['dim'] // 2
+            up = general_params['U'][icrsh] - 2.0 * general_params['J'][icrsh]
+            j = general_params['J'][icrsh]
+        h_int = ftps.solver_core.HInt(u=general_params['U'][icrsh], j=j, up=up, dd=False)
+    elif sum_k.SO == 0:
+        # Constructs U matrix
+        Umat, Upmat = util.U_matrix_kanamori(n_orb=n_orb, U_int=general_params['U'][icrsh],
+                                             J_hund=general_params['J'][icrsh])
 
-        if n_orb not in (2, 3):
-            mpi.report('warning: are you sure you want to use the kanamori hamiltonian '
-                       + 'outside the t2g or eg manifold?')
-
-        if general_params['solver_type'] == 'ftps':
-            # 1-band modell requires J and U' equals zero
-            if n_orb == 1:
-                up, j = 0.0, 0.0
-            else:
-                up = general_params['U'][icrsh] - 2.0 * general_params['J'][icrsh]
-                j = general_params['J'][icrsh]
-            h_int[icrsh] = ftps.solver_core.HInt(u=general_params['U'][icrsh], j=j, up=up, dd=False)
-        elif sum_k.SO == 0:
-            # Constructs U matrix
-            Umat, Upmat = util.U_matrix_kanamori(n_orb=n_orb, U_int=general_params['U'][icrsh],
-                                                 J_hund=general_params['J'][icrsh])
-
-            h_int[icrsh] = util.h_int_kanamori(sum_k.spin_block_names[sum_k.SO], orb_names,
-                                               map_operator_structure=sum_k.sumk_to_solver[icrsh],
-                                               U=Umat, Uprime=Upmat, J_hund=general_params['J'][icrsh],
-                                               H_dump=os.path.join(general_params['jobname'], 'H.txt'))
-        else:
[…]
