### 2022-08-16 b199903
add Ntot interaction
