### 2022-08-16 683dda1
fix print for failing comparison test
