### 2022-08-17 617e696
add pick solver struct option to select specific blocks for the impurity problem
