+++ python/solid_dmft/dmft_tools/observables.py
@@ -423,8 +423,13 @@ def write_obs(observables, sum_k, general_params):
 
     """
 
-    n_orbitals = [sum_k.corr_shells[sum_k.inequiv_to_corr[iineq]]['dim']
-                  for iineq in range(sum_k.n_inequiv_shells)]
+    n_orbitals = [{'up': 0, 'down': 0}] * sum_k.n_inequiv_shells
+    for icrsh in range(sum_k.n_inequiv_shells):
+        for block, n_orb in sum_k.gf_struct_solver[icrsh].items():
+            if 'down' in block:
+                n_orbitals[icrsh]['down'] += sum_k.gf_struct_solver[icrsh][block]
+            else:
+                n_orbitals[icrsh]['up'] += sum_k.gf_struct_solver[icrsh][block]
 
     for icrsh in range(sum_k.n_inequiv_shells):
         if not general_params['csc'] and general_params['magnetic'] and sum_k.SO == 0:
@@ -432,13 +437,13 @@ def write_obs(observables, sum_k, general_params):
                 line = '{:3d} | '.format(observables['iteration'][-1])
                 line += '{:10.5f} | '.format(observables['mu'][-1])
 
-                if n_orbitals[icrsh] == 1:
+                if n_orbitals[icrsh][spin] == 1:
                     line += ' '*11
                 for item in observables['orb_gb2'][icrsh][spin][-1]:
                     line += '{:10.5f}   '.format(item)
                 line = line[:-3] + ' | '
 
-                if n_orbitals[icrsh] == 1:
+                if n_orbitals[icrsh][spin] == 1:
                     line += ' '*11
                 for item in observables['orb_occ'][icrsh][spin][-1]:
                     line += '{:10.5f}   '.format(item)
@@ -461,19 +466,19 @@ def write_obs(observables, sum_k, general_params):
             line += '{:10.5f} | '.format(observables['mu'][-1])
 
             # Adds spaces for header to fit in properly
-            if n_orbitals[icrsh] == 1:
+            if n_orbitals[icrsh]['up'] == 1:
                 line += ' '*11
             # Adds up the spin channels
-            for iorb in range(n_orbitals[icrsh]):
+            for iorb in range(n_orbitals[icrsh]['up']):
                 val = np.sum([observables['orb_gb2'][icrsh][spin][-1][iorb] for spin in sum_k.spin_block_names[sum_k.SO]])
                 line += '{:10.5f}   '.format(val)
             line = line[:-3] + ' | '
 
             # Adds spaces for header to fit in properly
-            if n_orbitals[icrsh] == 1:
+            if n_orbitals[icrsh]['up'] == 1:
                 line += ' '*11
             # Adds up the spin channels
-            for iorb in range(n_orbitals[icrsh]):
+            for iorb in range(n_orbitals[icrsh]['up']):
                 val = np.sum([observables['orb_occ'][icrsh][spin][-1][iorb] for spin in sum_k.spin_block_names[sum_k.SO]])
                 line += '{:10.5f}   '.format(val)
             line = line[:-3] + ' | '
