### 2022-08-17 b24b3a8
fix header writing for obs file
