### 2022-08-17 f040765
aplly block picker before block mapping
