### 2022-08-19 ea65cf6
function to det n_orb_solver, fix test
