+++ python/solid_dmft/dmft_tools/solver.py
@@ -33,6 +33,30 @@ from h5 import HDFArchive
 
 from . import legendre_filter
 
+def n_orb_solver(sum_k):
+    """
+    determines the number of orbitals within the
+    solver block structure.
+
+    Parameters
+    ----------
+    sum_k : dft_tools sumk object
+
+    Returns
+    -------
+    n_orb : dict of int
+        number of orbitals for up / down as dict
+    """
+    n_orbitals = [{'up': 0, 'down': 0} for i in range(sum_k.n_inequiv_shells)]
+    for icrsh in range(sum_k.n_inequiv_shells):
+        for block, n_orb in sum_k.gf_struct_solver[icrsh].items():
+            if 'down' in block:
+                n_orbitals[icrsh]['down'] += sum_k.gf_struct_solver[icrsh][block]
+            else:
+                n_orbitals[icrsh]['up'] += sum_k.gf_struct_solver[icrsh][block]
+
+    return n_orbitals
+
 def _gf_fit_tail_fraction(Gf, fraction=0.4, replace=None, known_moments=None):
     """
     fits the tail of Gf object by making a polynomial
