### 2022-08-22 bbd68d8
adapt vasp patch files for ver6.3.2
