### 2022-08-22 fd41678
adapt vasp patch files for ver6.3.2
