### 2022-08-23 #4623 @amandadumi
Hello, checking in to say I am summarizing my thoughts on the manuscript and will make progress on the code review in the following days. I had an event that needed preparation that was dictating my time, but that’s over now!
