### 2022-08-25 flatironinstitute/solid_dmft#20 @shivupa
"Seeking Help" section in docs
### Summary
Part of Review: https://github.com/openjournals/joss-reviews/issues/4623

The "How to contribute" is great, and could a "Seeking Help" section be added to the docs?

### Motivation

This would make it easier if users encounter issues.

### Implementation
I'd expect this to point at the tutorials/examples, to link to the Github Discussions page, and possibly link to opening an issue.
