### 2022-08-25 flatironinstitute/solid_dmft#21 @shivupa
Review suggestion: Build script block at end of building from source section
### Summary
[Part of Joss Review](https://github.com/openjournals/joss-reviews/issues/4623)

This tool fits very nicely within the Triqs ecosystem and its existence makes sense/does not seem redundant. The installation of solid-dmft follows a similar flow as the other tools within the ecosystem, which is nice from a user's perspective. One thing that would help slightly is a collected build script for the build from source at the end of the installation from source section. [Triqs](https://triqs.github.io/triqs/latest/install.html#compiling-triqs-from-source-advanced) has one in the installation instructions which is little more than a collection of the instructions that preceed it, but it makes copy+pasting easier.
