+++ python/solid_dmft/postprocessing/plot_correlated_bands.py
@@ -45,8 +45,9 @@ import itertools
 import skimage.measure
 
 from h5 import HDFArchive
-from triqs.gf import BlockGf, MeshReFreq
+from triqs.gf import BlockGf, MeshReFreq, Gf
 from triqs.lattice.utils import TB_from_wannier90, k_space_path
+from triqs_dft_tools.sumk_dft import SumkDFT
 
 
 def _linefit(x, y, interval, spacing=50, addspace=0.0):
@@ -107,20 +108,45 @@ def print_matrix(matrix, n_orb, text):
 
 def _sigma_from_dmft(n_orb, orbital_order, with_sigma, spin, block, orbital_order_dmft, eta=0.0, **specs):
 
+    block_spin = spin + '_' + str(block) if with_sigma == 'calc' else spin
+
     if with_sigma == 'calc':
         print('Setting Sigma from {}'.format(specs['dmft_path']))
 
+        sigma_imp_list = []
+        dc_list = []
         with HDFArchive(specs['dmft_path'], 'r') as ar:
-            try:
-                sigma = ar['DMFT_results'][specs['it']]['Sigma_freq_0']
-                assert isinstance(sigma.mesh, MeshReFreq), 'Imported Greens function must be real frequency'
-            except(KeyError, AssertionError):
+            for icrsh in range(ar['dft_input']['n_inequiv_shells']):
                 try:
-                    sigma = ar['DMFT_results'][specs['it']]['Sigma_maxent_0']
-                except KeyError:
-                    raise KeyError('Provide either "Sigma_freq_0" in real frequency or "Sigma_maxent_0".')
-            dc = ar['DMFT_results'][specs['it']]['DC_pot'][0][spin][0, 0]
+                    sigma = ar['DMFT_results'][specs['it']][f'Sigma_freq_{icrsh}']
+                    assert isinstance(sigma.mesh, MeshReFreq), 'Imported Greens function must be real frequency'
+                except(KeyError, AssertionError):
+                    try:
+                        sigma = ar['DMFT_results'][specs['it']][f'Sigma_maxent_{icrsh}']
+                    except KeyError:
+                        try:
+                            sigma = ar['DMFT_results'][specs['it']][f'Sigma_Refreq_{icrsh}']
+                        except KeyError:
+                            raise KeyError('Provide either "Sigma_freq_0" in real frequency, "Sigma_Refreq_0" or "Sigma_maxent_0".')
+                dc = ar['DMFT_results'][specs['it']]['DC_pot'][icrsh][spin]
+
+                sigma_imp_list.append(sigma - dc)
+
             mu_dmft = ar['DMFT_results'][specs['it']]['chemical_potential_post']
+
+            sum_k = SumkDFT(specs['dmft_path'], mesh=sigma.mesh)
+            sum_k.block_structure = ar['DMFT_input/block_structure']
+            sum_k.deg_shells = ar['DMFT_input/deg_shells']
+            sum_k.set_mu = mu_dmft
+
+            sigma_sumk = sum_k.transform_to_sumk_blocks(sigma_imp_list)
+
+            sigma = Gf(mesh=sigma.mesh, target_shape=[n_orb, n_orb])
+            for icrsh in range(ar['dft_input']['n_corr_shells']):
+                sigma += sum_k.upfold(ik=0, ish=icrsh, bname=spin, gf_to_upfold=sigma_sumk[icrsh][spin], gf_inp=sigma)
+
+        # already subtracted
+        dc = 0.0
 
     else:
[…]
