### 2022-09-12 62fb3e3
DOC: update support & contribute section, bump ver to 3.1.1
