### 2022-09-12 flatironinstitute/solid_dmft#20 @the-hampel
Hi @shivupa,
thank you for your feedback. We incorporated your suggestions in commit 62fb3e397bd2516e09070d2abd5ae3167bb63a4a and the changes are online: https://triqs.github.io/solid_dmft/issues.html . We decided to slightly rename the section to `Support & contribute` and then add a subsection `seeking help` . We also revised the text to point to the github discussion page, issues, and the tutorials.
