### 2022-09-13 ede7a2d
Suggestion for addressing issue 23 and 24
