+++ paper/paper.md
@@ -41,36 +41,35 @@ Given the complexity of the method, there is also the question of whether users 
 The goal of solid\_dmft is to find a middle ground, i.e., a *gray-box* tool as a ready-to-use implementation.
 This means that while the code contains all the functionality needed for many standard DMFT calculations, it is highly modular, based on open-source and community-developed software, and therefore can be easily adapted to specific applications and needs.
 Hence, this project is targeted towards researchers aiming to apply DMFT methods on top of DFT simulations to describe the physics of strongly correlated electron systems.
-While our approach allows one to fully perform these computations using standardized input flags without further need for elaborate coding, the final user can easily extend the functionalities by modifying the corresponding modules in the code.
+While our approach allows one to fully perform these computations using standardized input flags without need for coding, the final user can easily extend the functionalities by modifying the corresponding modules in the code.
 
-The package is MPI-parallelized and written in Python 3, utilizing the publicly available TRIQS software library [@triqs:2015] and its available applications.
-The philosophy of the package is to increase reproducibility of DFT+DMFT calculations, provide clearer convergence metrics, and allow to run calculations for a large variety of systems without adapting the code manually, i.e., on a level similar to widely available DFT simulation packages.
+The package is MPI-parallelized and written in Python 3, utilizing the publicly available TRIQS software library [@triqs:2015] and the applications based on TRIQS, such as different solvers or DFT interfaces.
+The philosophy of the package is to increase reproducibility of DFT+DMFT calculations, provide clearer convergence metrics, and allow one to run calculations for a large variety of systems without adapting the code manually, i.e., on a level similar to widely available DFT simulation packages.
 
 ## Design Principles
 
-solid_dmft uses the state-of-the-art implementations provided by the TRIQS library and its applications.
-This allows to easily run ab-initio calculations for strongly correlated materials, as well as implement and test new features of TRIQS and benchmark new solvers against existing ones.
-solid\_dmft manages the calls of the necessary routines to run the DFT calculations, create the downfolded Hamiltonian, solve the resulting Hubbard-like Hamiltonian via DMFT, postprocess the data to calculate physical observables, and allow for charge-corrected feedback via charge self-consistency.
+solid_dmft uses the state-of-the-art implementations provided by the TRIQS ecosystem.
+This allows the user to easily run ab-initio calculations for strongly correlated materials, as well as implement and test new features of TRIQS and benchmark new TRIQS solvers against existing ones.
+solid\_dmft either manages the DFT run itself in charge self-consistency or simply uses the DFT output for one-shot calculations, then creates the downfolded Hamiltonian, solves the resulting Hubbard-like Hamiltonian via DMFT, postprocesses the data to calculate physical observables, and allows feeding back the corrected charge via charge self-consistency.
 The full DFT+DMFT cycle is presented in \autoref{fig:downfolding}.
 
 ![Fully charge self-consistent DFT+DMFT cycle. Starting from a DFT calculation (top left), a downfolded Hamiltonian and projector functions are created via optimized projections on a local basis set (top right). By adding a specified interaction Hamiltonian $H_\text{int}$, a full interacting electron problem is created, to be solved via the DMFT equations in TRIQS (bottom). After convergence in DMFT is reached, physical observables are calculated (bottom left). For fully charge self-consistent calculations, the DFT cycle is restarted with a DMFT-corrected charge density.\label{fig:downfolding}](downfolding.png){ width=100% }
 
-The code is designed to run on top of a DFT calculation or a tight-binding model, which provide the low-energy (downfolded) description of a periodic system.
[…]
