### 2022-09-13 flatironinstitute/solid_dmft#20 @shivupa
Excellent thanks! Closing as completed
