+++ paper/paper.bib
@@ -264,3 +264,26 @@ eprint = {https://royalsocietypublishing.org/doi/pdf/10.1098/rspa.1963.0204}
 	year = {1996},
 	pages = {11169--11186},
 }
+
+@article{sun_pyscf_2018,
+	title = {{PySCF}: the {Python}-based simulations of chemistry framework},
+	volume = {8},
+	doi = {10.1002/wcms.1340},
+	number = {1},
+	journal = {WIREs Computational Molecular Science},
+	author = {Sun, Qiming and Berkelbach, Timothy C. and Blunt, Nick S. and Booth, George H. and Guo, Sheng and Li, Zhendong and Liu, Junzi and McClain, James D. and Sayfutyarova, Elvira R. and Sharma, Sandeep and Wouters, Sebastian and Chan, Garnet Kin-Lic},
+	year = {2018},
+	pages = {e1340},
+}
+
+@article{larsen_atomic_2017,
+	title = {The atomic simulation environment—a {Python} library for working with atoms},
+	volume = {29},
+	doi = {10.1088/1361-648X/aa680e},
+	number = {27},
+	journal = {Journal of Physics: Condensed Matter},
+	author = {Larsen, Ask Hjorth and Mortensen, Jens Jørgen and Blomqvist, Jakob and Castelli, Ivano E. and Christensen, Rune and Du{\textbackslash}lak, Marcin and Friis, Jesper and Groves, Michael N. and Hammer, Bjørk and Hargus, Cory and Hermes, Eric D. and Jennings, Paul C. and Jensen, Peter Bjerre and Kermode, James and Kitchin, John R. and Kolsbjerg, Esben Leonhard and Kubal, Joseph and Kaasbjerg, Kristen and Lysgaard, Steen and Maronsson, Jón Bergmann and Maxson, Tristan and Olsen, Thomas and Pastewka, Lars and Peterson, Andrew and Rostgaard, Carsten and Schiøtz, Jakob and Schütt, Ole and Strange, Mikkel and Thygesen, Kristian S. and Vegge, Tejs and Vilhelmsen, Lasse and Walter, Michael and Zeng, Zhenhua and Jacobsen, Karsten W.},
+	year = {2017},
+	pages = {273002},
+}
+
