+++ paper/paper.md
@@ -38,7 +38,7 @@ In these systems, an effective single-particle description may not capture the m
 Although density functional theory (DFT) plus dynamical mean-field theory (DMFT) has proven successful in describing strongly correlated electron systems for over two decades, it has only been very recently that ready-to-use software packages began to become available, with most scientific research carried out by self-written codes developed and used in research groups.
 Given the complexity of the method, there is also the question of whether users should implement the formalism themselves for each problem or whether ready-to-use black-box software, as is the case with many DFT software packages, is beneficial to the community.
 
-The goal of solid\_dmft is to find a middle ground, i.e., a *gray-box* tool as a ready-to-use implementation. Such an approach is widely used in other areas of materials simulation (pyscf, ase).
+The goal of solid\_dmft is to find a middle ground, i.e., a *gray-box* tool as a ready-to-use implementation. Such a gray-box approach is widely used in other areas of materials simulation [@sun_pyscf_2018; @larsen_atomic_2017].
 This means that while the code contains all the functionality needed for many standard DMFT calculations, it is highly modular, based on open-source and community-developed software, and therefore can be easily adapted to specific applications and needs.
 Hence, this project is targeted towards researchers aiming to apply DMFT methods on top of DFT simulations to describe the physics of strongly correlated electron systems.
 While our approach allows one to fully perform these computations using standardized input flags without need for coding, the final user can easily extend the functionalities by modifying the corresponding modules in the code.
@@ -50,7 +50,7 @@ The philosophy of the package is to increase reproducibility of DFT+DMFT calcula
 
 solid_dmft uses the state-of-the-art implementations provided by the TRIQS ecosystem.
 This allows the user to easily run ab-initio calculations for strongly correlated materials, as well as implement and test new features of TRIQS and benchmark new TRIQS solvers against existing ones.
-solid\_dmft either manages the DFT run itself in the charge-self-consistent mode or simply postprocesses the DFT output for one-shot calculations. In both cases, solid\_dmft creates the downfolded Hamiltonian, solves the resulting Hubbard-like Hamiltonian via DMFT, calculates physical observables from the DMFT data, and in case of charge self-consistency feeds back the corrected charge to DFT.
+solid\_dmft either manages the DFT run itself in the charge-self-consistent mode or simply postprocesses the DFT output for one-shot calculations. In both cases, solid\_dmft then creates the downfolded Hamiltonian, solves the resulting Hubbard-like Hamiltonian via DMFT, calculates physical observables from the DMFT data, and in case of charge self-consistency feeds back the corrected charge to DFT.
 The full DFT+DMFT cycle is presented in \autoref{fig:downfolding}.
 
 ![Fully charge self-consistent DFT+DMFT cycle. Starting from a DFT calculation (top left), a downfolded Hamiltonian and projector functions are created via optimized projections on a local basis set (top right). By adding a specified interaction Hamiltonian $H_\text{int}$, a full interacting electron problem is created, to be solved via the DMFT equations in TRIQS (bottom). After convergence in DMFT is reached, physical observables are calculated (bottom left). For fully charge self-consistent calculations, the DFT cycle is restarted with a DMFT-corrected charge density.\label{fig:downfolding}](downfolding.png){ width=100% }
