### 2022-09-14 flatironinstitute/solid_dmft#24 @merkelm
Thank you for the feedback! We changed the paper accordingly in the pull request #26 on the `joss` branch, together with a few changes related to issue #23. The only bullet points of yours we did not address by changing the paper are the following (numbers referring to the line numbers you used in this issue):

- 46: We construct the localized orbitals from the single-particle Kohn-Sham wavefunction from DFT. This can be achieved by either projecting on localized atomic-like orbitals or more elaborate wannierization procedures.
- 49/62: According to [Wikipedia](https://en.wikipedia.org/wiki/Green%27s_function_(many-body_theory)), both Green's function and Green function are acceptable. We decided to go for Green function here to be consistent with other terms, such as Hubbard U, Lorentz transformation, etc
- 65: For 2), we stuck to the term "reference manual". The difference is that a documentation provides higher-level overview while a reference manual documents the behavior of functions or classes in the source code. We updated the paper following your points 1) and 3).
