### 2022-09-14 945373b
DOC: extend install instructions & improve readme for #21 #22
