### 2022-09-14 b0a18f1
DOC: extend install instructions & improve readme for #21 #22
