### 2022-09-14 74fd4c7
DOC: extend install instructions & improve readme for #21 #22
