### 2022-09-14 flatironinstitute/solid_dmft#21 @the-hampel
Hi @shivupa,
we followed your advice and made a copy/paste friendly version of the install instructions: https://triqs.github.io/solid_dmft/install.html . We also link to the docker files, which hold also full installation scripts in an extra section.
Best,
Alex
