### 2022-09-15 aef7cbe
Few small bug fixes
