+++ python/solid_dmft/dmft_tools/observables.py
@@ -40,7 +40,7 @@ import triqs.utility.mpi as mpi
 from triqs.gf import Gf, MeshImTime
 from triqs.atom_diag import trace_rho_op
 from triqs.gf.descriptors import Fourier
-from solid_dmft.dmft_tools.solver import n_orb_solver
+from solid_dmft.dmft_tools import solver
 
 def prep_observables(h5_archive, sum_k):
     """
@@ -97,7 +97,7 @@ def _generate_header(general_params, sum_k):
     Generates the headers that are used in write_header_to_file.
     Returns a dict with {file_name: header_string}
     """
-    n_orb = n_orb_solver(sum_k)
+    n_orb = solver.get_n_orbitals(sum_k)
 
     header_energy_mask = ' | {:>10} | {:>10}   {:>10}   {:>10}   {:>10}'
     header_energy = header_energy_mask.format('E_tot', 'E_DFT', 'E_bandcorr', 'E_int_imp', 'E_DC')
@@ -423,7 +423,7 @@ def write_obs(observables, sum_k, general_params):
 
     """
 
-    n_orb = n_orb_solver(sum_k)
+    n_orb = solver.get_n_orbitals(sum_k)
 
     for icrsh in range(sum_k.n_inequiv_shells):
         if not general_params['csc'] and general_params['magnetic'] and sum_k.SO == 0:
