### 2022-09-16 8cb2a6d
include suggestions of @merkelm
* improve doc string of pick and map gf_struct_solver
* add valid_for condition for ratio_F4_F2
* map_gf_struct_solver must now be a list for each imp
* check that pick and map_gf_struct_solver are a list
