### 2022-09-16 babd79d
fix parameter n_orb in hint construction
when pick_gf_struct_solver or map_gf_struct solver is used, the number
of orbital n_orb of sum_k.corr_shell[ish].dim is different from the
actual solver space dimension. Use function get_n_orbitals to determine
correct number of orbitals.
