### 2022-09-16 #4623 @the-hampel
Hi, this is just to let you know that we addressed all open issues and corrected the paper (joss branch) and corresponding parts in the code on github (unstable branch).
- Alex
