### 2022-09-19 flatironinstitute/solid_dmft#21 @shivupa
Nice happy to see the dockerfiles as well!
