### 2022-09-19 #4623 @shivupa
OK Sorry for the speed of this review, but I just wanted to say that the issues I have opened have been addressed.

I had worked through one of the tutorials (SrVO3 Metal Insulator transition), and I was able to get a little bit of experience of using the code. The example does require pretty significant computational resources, but this was mentioned in the tutorial so it wasn't a surprise to me. I think having real world examples as the tutorial rather than extremely lightweight/minimal tutorials is preferable in my eyes.

I do recommend this for acceptance (once the unstable branch is merged), and it does seem like this will make DMFT calculations more accessible.
