### 2022-09-19 45fe8b1
fix deg shells in magnetic calculations
the old method of removing degenerate shells in magnetic calculation was
faulty and could sometimes not correctly remove a degeneracy between up
and down orbitals. The completely revised version does the job correctly
and removes all degeneracies in case of SOC calculations, with
magnetism.
