### 2022-09-20 flatironinstitute/solid_dmft#24 @amandadumi
Thank you for the update. The use of reference manual makes sense, the context on the use of "wave functions" in your first point gives the context I needed. I had never seen the phrase Green function used so thanks for clarifying, I'll keep that in mind moving forward.

This all looks good and comments have all addressed, thank you!
