### 2022-09-20 #4623 @amandadumi
Hello,

I also want to apologize that this took me longer than intended, but thanks for the patience!

I was able to work through a different tutorial, "3. CSC with QE/W90 and HubbardI: total energy in Ce2O". This specific tutorial was reasonable in computational cost and was easy to work through. A quick look at the other tutorials suggests that these are also seemed informative and relevant to real application. I agree with @shivupa that the increased cost is offset by the insight gained. The code and documentation and use was also clearly organized and structured, with all needed information for beginning users present.

The community benefit is significant as the code is currently written, but also in the modular design and well-described "how to contribute" section. This encourages the transition from isolated in-house workflows of other groups into a more transferable and sustainable form, which will benefit the broader community who wish to use or develop DMFT approaches as well.

I do recommend this work for publication in JOSS!
