### 2022-09-20 94a458c
Bug fix for changes in sumk mesh definition in maxent_gf_latt
