### 2022-09-20 #4623 @rkurchin
Some minor editorial comments:
* line 14: should read "...packages *have begun* to become..." (tense consistency)
* 15: "self-written" is a bit of an awkward phrase. I might rephrase this to something like "...carried out by in-house codes developed and used in individual research groups."
* 17: a bit awkward (and strictly ungrammatical) as-written here, maybe rephrase to "...whether black-box software, such as is available for DFT, is beneficial to the community" or something like "...whether black-box software, analogous to popular DFT packages, would be beneficial..."
* 27: "end user" is a more standard phrase than "final user"
* 27-28: suggest rephrase to "...can easily extend the functionality by modifying relevant modules in the code."
* 30: delete "the" from "...and the applications based on TRIQS..."
* 31: I would use the word "goal" instead of "philosophy" here, since that's more what's being expressed
* 39 + caption of figure 1: be consistent about whether the first hyphen is present or not in "charge-self-consistent" (I lean towards including it)
* 48: "reciprocal k-space" is redundant...either remove "reciprocal" or "k-" or if you really want both, put one in parentheses e.g. "reciprocal (k-)space"
* 51, 63: should read "Green's functions"
* 52: "limit statefulness to a minimum" is redundant...either replace "limit" with "keep" or remove "to a minimum"
* 71: should read "only a few" or "very few"
* 76-77: "open-source" should be hyphenated
