+++ paper/paper.md
@@ -35,16 +35,16 @@ bibliography: paper.bib
 
 Strongly correlated systems are a class of materials whose electronic structure is heavily influenced by the effect of electron-electron interactions.
 In these systems, an effective single-particle description may not capture the many-body effects accurately.
-Although density functional theory (DFT) plus dynamical mean-field theory (DMFT) has proven successful in describing strongly correlated electron systems for over two decades, it has only been very recently that ready-to-use software packages began to become available, with most scientific research carried out by self-written codes developed and used in research groups.
-Given the complexity of the method, there is also the question of whether users should implement the formalism themselves for each problem or whether ready-to-use black-box software, as is the case with many DFT software packages, is beneficial to the community.
+Although density functional theory (DFT) plus dynamical mean-field theory (DMFT) has proven successful in describing strongly correlated electron systems for over two decades, only very recently ready-to-use software packages have begun to become available, with most scientific research carried out by in-house codes developed and used in individual research groups.
+Given the complexity of the method, there is also the question of whether users should implement the formalism themselves for each problem or whether black-box software, analogous to popular DFT packages, would be beneficial to the community.
 
 The goal of solid\_dmft is to find a middle ground, i.e., a *gray-box* tool as a ready-to-use implementation. Such a gray-box approach is widely used in other areas of materials simulation [@sun_pyscf_2018; @larsen_atomic_2017].
 This means that while the code contains all the functionality needed for many standard DMFT calculations, it is highly modular, based on open-source and community-developed software, and therefore can be easily adapted to specific applications and needs.
 Hence, this project is targeted towards researchers aiming to apply DMFT methods on top of DFT simulations to describe the physics of strongly correlated electron systems.
-While our approach allows one to fully perform these computations using standardized input flags without need for coding, the final user can easily extend the functionalities by modifying the corresponding modules in the code.
+While our approach allows one to fully perform these computations using standardized input flags without need for coding, the end user can easily extend the functionality by modifying relevant modules in the code.
 
-The package is MPI-parallelized and written in Python 3, utilizing the publicly available TRIQS software library [@triqs:2015] and the applications based on TRIQS, such as different solvers or interfaces to DFT codes.
-The philosophy of the package is to increase reproducibility of DFT+DMFT calculations, provide clearer convergence metrics, and allow one to run calculations for a large variety of systems without adapting the code manually, i.e., on a level similar to widely available DFT simulation packages.
+The package is MPI-parallelized and written in Python 3, utilizing the publicly available TRIQS software library [@triqs:2015] and applications based on TRIQS, such as different solvers or interfaces to DFT codes.
+The goal of this package is to increase reproducibility of DFT+DMFT calculations, provide clearer convergence metrics, and allow one to run calculations for a large variety of systems without adapting the code manually, i.e., on a level similar to widely available DFT simulation packages.
 
 ## Design Principles
 
@@ -53,14 +53,14 @@ This allows the user to easily run ab-initio calculations for strongly correlate
 solid\_dmft either manages the DFT run itself in the charge-self-consistent mode or simply postprocesses the DFT output for one-shot calculations. In both cases, solid\_dmft then creates the downfolded Hamiltonian, solves the resulting Hubbard-like Hamiltonian via DMFT, calculates physical observables from the DMFT data, and in case of charge self-consistency feeds back the corrected charge to DFT.
 The full DFT+DMFT cycle is presented in \autoref{fig:downfolding}.
 
[…]
