### 2022-09-21 #4623 @merkelm
Dear @rkurchin, thank you for your comments on the paper! I just incorporated them in [this commit](https://github.com/TRIQS/solid_dmft/commit/3ec8bc363352bb3048dd600350bc7a2b38b0f470).
-- Max
