### 2022-09-21 #4623 @rkurchin
Great! So once you finish up the stuff in [this comment](https://github.com/openjournals/joss-reviews/issues/4623#issuecomment-1252741234) we should be good to go!
