### 2022-09-21 df36d42
bump version number to 3.1.2
