+++ packaging/conda/meta.yaml
@@ -1,4 +1,4 @@
-{% set version = "3.0.0" %}
+{% set version = "3.1.2" %}
 
 package:
   name: solid_dmft
