+++ CMakeLists.txt
@@ -30,7 +30,7 @@ endif()
 
 # ############
 # Define Project
-project(solid_dmft VERSION 3.1.1 LANGUAGES C CXX)
+project(solid_dmft VERSION 3.1.2 LANGUAGES C CXX)
 get_directory_property(IS_SUBPROJECT PARENT_DIRECTORY)
 
 # ############
