### 2022-09-22 #4623 @the-hampel
Dear @rkurchin,
I hope I did everything correctly: https://doi.org/10.5281/zenodo.7105085 . This Zenodo DOI refers to a new release 3.1.2 : https://github.com/TRIQS/solid_dmft/releases/tag/3.1.2 . I included in this release all changes we made during the review process plus some small fixes to the code. We decided to keep the JOSS branch separate, so that people download the paper from the JOSS webpage. Let me know if there are any problems with it.
