### 2022-09-23 #4623 @kyleniemeyer
Congratulations @the-hampel on your article's publication in JOSS!8614))
