### 2022-09-23 #4623 @the-hampel
Hi,
I just noticed that we included a back-slash in the title in the markdown file, which is weirdly shown in the title: `solid\_dmft` . In the pdf this looks fine. Is there still a chance to change this? Sorry for the late notice, but in the pdf everything is fine. It just shown on the joss webpage.
