### 2022-09-23 #4623 @kyleniemeyer
> Hi,
> I just noticed that we included a back-slash in the title in the markdown file, which is weirdly shown in the title on the webpage: `solid\_dmft` . In the pdf this looks fine. Is there still a chance to change this? Sorry for the late notice, but in the pdf everything is fine. It just shown on the joss webpage. (edit: just noticed, that I was 10min too slow)

@openjournals/dev do you know how to fix this issue?
