### 2022-09-25 #4623 @the-hampel
Hi @arfon ,
I removed all backslashes in the markdown: https://github.com/TRIQS/solid_dmft/commit/6d10d49d0c19529146b3609b593490f2632bc37d on our JOSS branch. No other changes. If you could re-process it would be awesome. Thank you!
